Ice contacts Appointment refactoring

+++ app/src/main/java/sg/edu/nus/iss/se/ft05/medipal/dao/AppointmentDAO.java
@@ -3,18 +3,26 @@ package sg.edu.nus.iss.se.ft05.medipal.dao;
 /**
  * Created by Dhruv on 18/3/2017.
  */
+
 import android.database.Cursor;
-import android.util.Log;
 
 import java.util.List;
 
 import sg.edu.nus.iss.se.ft05.medipal.model.Appointment;
+
 public interface AppointmentDAO {
-    public int delete(long id);
+
+    public int delete(int id);
+
     public Cursor findAll();
-    public Appointment findById(long id);
+
+    public Appointment findById(int id);
+
     public long insert(Appointment appointment);
+
     public int update(Appointment appointment);
+
     public List<Appointment> findByDate(String date);
+
     public Cursor filterDate(String date);
 }
